Merged 4 different script type directories into a single directory. The following 4 directories:
 * ambra/libs/article-util
 * ambra/libs/runscripts
 * ambra/libs/scripts
 * ambra/libs/util

have been merged into a single directory ambra/scripts. This was mostly done
using git-mv so hopefully the historic links have been maintained.

There is one catch here that I am not happy with and need to think a little
more. annotated.groovy and alertsinfo.groovy in ambra-tools rpm appear in two
places. It is not a big deal when number of files are small, but this still
needs to be cleaned up. This files can be compiled and not made part of the
package as text files is an option.