Upgrade OWL generation code to somewhat work again. There is a problem with the gmaven-plugin classloader, so unfortunately this will not work using the old: {{{ mvn -Pgenerate-owl -Dargs="..." }}}
gmaven-plugin is setting the wrong class context loader for threads causing
problem for OTM to load the id generation classes. If you want this to work in
your work directory currently, change the loading of id generation in
AnnotationClassMetaFactory to getClass.getClassLoader() instead of
Thread.currentThread().getContextClassLoader(). DO NOT CHECK THAT CODE IN
HOWEVER.

There is still one major issue with OWLGenerator related to properties in
non-entity marked super classes that is pending. That will help reduce the
duplication of properties in various sibling classes.